Renames test_cases to ocean_only and clean Eady_channel files
