+++ ocean_only/Eady_channel/MOM_override
@@ -1,0 +1,1 @@
+! Blank file in which we can put "overrides" for parameters
