--- test_cases/Eady_channel/MOM_parameter_doc.layout
@@ -1,73 +1,0 @@
-! This file was written by the model and records the layout parameters used at run-time.
-GLOBAL_INDEXING = False         !   [Boolean] default = False
-                                ! If true, use a global lateral indexing convention, so
-                                ! that corresponding points on different processors have
-                                ! the same index. This does not work with static memory.
-!SYMMETRIC_MEMORY_ = False      !   [Boolean]
-                                ! If defined, the velocity point data domain includes
-                                ! every face of the thickness points. In other words,
-                                ! some arrays are larger than others, depending on where
-                                ! they are on the staggered grid.  Also, the starting
-                                ! index of the velocity-point arrays is usually 0, not 1.
-                                ! This can only be set at compile time.
-NONBLOCKING_UPDATES = False     !   [Boolean] default = False
-                                ! If true, non-blocking halo updates may be used.
-THIN_HALO_UPDATES = False       !   [Boolean] default = False
-                                ! If true, optional arguments may be used to specify the
-                                ! The width of the halos that are updated with each call.
-!STATIC_MEMORY_ = False         !   [Boolean]
-                                ! If STATIC_MEMORY_ is defined, the principle variables
-                                ! will have sizes that are statically determined at
-                                ! compile time.  Otherwise the sizes are not determined
-                                ! until run time. The STATIC option is substantially
-                                ! faster, but does not allow the PE count to be changed
-                                ! at run time.  This can only be set at compile time.
-NIHALO = 4                      ! default = 4
-                                ! The number of halo points on each side in the
-                                ! x-direction.  With STATIC_MEMORY_ this is set as NIHALO_
-                                ! in MOM_memory.h at compile time; without STATIC_MEMORY_
-                                ! the default is NIHALO_ in MOM_memory.h (if defined) or 2.
-NJHALO = 4                      ! default = 4
-                                ! The number of halo points on each side in the
-                                ! y-direction.  With STATIC_MEMORY_ this is set as NJHALO_
-                                ! in MOM_memory.h at compile time; without STATIC_MEMORY_
-                                ! the default is NJHALO_ in MOM_memory.h (if defined) or 2.
-MASKTABLE = "MOM_mask_table"    ! default = "MOM_mask_table"
-                                ! A text file to specify n_mask, layout and mask_list.
-                                ! This feature masks out processors that contain only land points.
-                                ! The first line of mask_table is the number of regions to be masked out.
-                                ! The second line is the layout of the model and must be
-                                ! consistent with the actual model layout.
-                                ! The following (n_mask) lines give the logical positions
-                                ! of the processors that are masked out. The mask_table
-                                ! can be created by tools like check_mask. The
-                                ! following example of mask_table masks out 2 processors,
-                                ! (1,2) and (3,6), out of the 24 in a 4x6 layout:
-                                !  2
-                                !  4,6
-                                !  1,2
-                                !  3,6
-NIPROC = 4                      !
-                                ! The number of processors in the x-direction. With
-                                ! STATIC_MEMORY_ this is set in MOM_memory.h at compile time.
-NJPROC = 4                      !
-                                ! The number of processors in the x-direction. With
-                                ! STATIC_MEMORY_ this is set in MOM_memory.h at compile time.
-LAYOUT = 4, 4                   !
-                                ! The processor layout that was acutally used.
-IO_LAYOUT = 1, 1                ! default = 1
-                                ! The processor layout to be used, or 0,0 to automatically
-                                ! set the io_layout to be the same as the layout.
-NIBLOCK = 1                     ! default = 1
-                                ! The number of blocks in the x-direction on each processor (for openmp).
-NJBLOCK = 1                     ! default = 1
-                                ! The number of blocks in the y-direction on each processor (for openmp).
-BT_USE_WIDE_HALOS = True        !   [Boolean] default = True
-                                ! If true, use wide halos and march in during the
-                                ! barotropic time stepping for efficiency.
-BTHALO = 0                      ! default = 0
-                                ! The minimum halo size for the barotropic solver.
-!BT x-halo = 0                  !
-                                ! The barotropic x-halo size that is actually used.
-!BT y-halo = 0                  !
-                                ! The barotropic y-halo size that is actually used.
