--- test_cases/Eady_channel/input.nml
@@ -1,33 +1,0 @@
- &MOM_input_nml
-         output_directory = './',
-         input_filename = 'n'
-         restart_input_dir = 'INPUT/',
-         restart_output_dir = 'RESTART/',
-         parameter_filename = 'MOM_input',
-                              'MOM_override'
-/
-
- &fms_io_nml
-            fms_netcdf_restart=.true.
-            threading_read='multi'
-            max_files_r = 200
-            max_files_w = 200
-/
-
- &fms_nml
-            clock_grain='ROUTINE'
-            clock_flags='NONE'
-            domains_stack_size = 960000
-            stack_size =0
-/
-
- &ocean_solo_nml
-            months = 0
-            days   = 1 
-            date_init = 1,1,1,0,0,0
-            hours = 0
-            minutes = 0
-            seconds = 0
-            calendar = 'NOLEAP'
-/
-
