--- test_cases/Eady_channel/timestats
@@ -1,10 +1,0 @@
-  Step,       Day,  Truncs,      Energy/Mass,      Maximum CFL,  Mean Sea Level,  Total Mass,  Mean Salin, Mean Temp, Frac Mass Err,   Salin Err,    Temp Err
-            [days]                 [m2 s-2]           [Nondim]       [m]             [kg]         [PSU]      [degC]       [Nondim]        [PSU]        [degC]
-     0,       0.000,     0, En 1.632312499997E-05, CFL  0.09286, SL  0.0000E+00, M 1.64317E+13, S  0.0000, T 16.9890, Me  0.00E+00, Se  0.00E+00, Te  0.00E+00
-    18,       0.250,     0, En 1.286258314142E-02, CFL  0.15628, SL  9.6634E-12, M 1.64317E+13, S  0.0000, T 16.9890, Me  4.82E-14, Se  0.00E+00, Te  8.20E-13
-    36,       0.500,     0, En 8.708665265013E-02, CFL  0.28017, SL  6.2499E-11, M 1.64317E+13, S  0.0000, T 16.9890, Me  2.64E-13, Se  2.99E-25, Te  4.49E-12
-    54,       0.750,     0, En 2.417846418540E-01, CFL  0.38094, SL  1.9904E-10, M 1.64317E+13, S  0.0000, T 16.9890, Me  6.83E-13, Se  1.74E-24, Te  1.16E-11
-    72,       1.000,     0, En 4.525136922608E-01, CFL  0.44959, SL  4.6947E-10, M 1.64317E+13, S -0.0000, T 16.9890, Me  1.35E-12, Se -8.18E-24, Te  2.30E-11
-    90,       1.250,   600, En 6.404474201671E-01, CFL  0.43778, SL  7.7836E-10, M 1.64317E+13, S -0.0000, T 16.9890, Me  1.54E-12, Se -8.58E-24, Te  2.63E-11
-   108,       1.500,  4000, En 7.818758408075E-01, CFL  0.41968, SL  1.0621E-09, M 1.64317E+13, S  0.0000, T 16.9890, Me  1.42E-12, Se  1.53E-23, Te  2.41E-11
-   126,       1.750,  7462, En 8.017032495076E-01, CFL  0.44474, SL  1.4704E-09, M 1.64317E+13, S  0.0000, T 16.9890, Me  2.04E-12, Se  1.00E-22, Te  3.47E-11
